started to refactor functionality to different classes. The most significant of which is flashcard_context, which encapsulates both the wordlist and sentences.
