+++ flashcards.py
@@ -1,18 +1,12 @@
 # -*- coding: utf-8 -*-
 #!/bin/python
 
-from unicode_csv import UnicodeReader 
 from export_loader import *
-import operator
-import random
+from entry import *
+from flashcard_context import *
+from sentence_filter import *
 
-minimum_desired_count = 1000
-maximum_card_count = 30000
-
-words = load_word_file("vocab.txt")
-extras = load_word_file("vocab_extras.txt")
-
-words = words + extras
+words = load_word_file("vocab.txt") + load_word_file("vocab_extras.txt")
 
 (sentences, unknown) = load_export_file("spoonfed_sample.txt", words)
 
@@ -22,96 +16,21 @@ words = words + extras
 # eliminate duplicates in sentence list (crudely)
 sentences = list(set(sentences))
 
-known_word_counts = dict()
-
-for s in sentences:
-	example = s.text
-
-	for w in words:
-		if w.text in example:
-			example = example.replace(w.text,"")
-			increment_word(known_word_counts, w)
-		
-			s.related.add(w)
-			w.related.add(s)
+context = Flashcard_Context(sentences, words)
 
 # Output list of words with no matching sentences
 #for w in words:
 #	if not w.text in known_word_counts:
 #		print w.text
 
-# null initialise added_word_count dictionary
-added_word_counts = dict()
 
-for w in known_word_counts:
-	added_word_counts[w] = 0
-
-covering = []
-added_sentences = set()
-
-looping = True
-
-while looping:
-	if len(covering) >= maximum_card_count or len(covering) >= len(sentences):
-		looping = False
-		break
-
-	# sort already added words incrementally by prevalence
-	it = sorted(added_word_counts.items(), key=operator.itemgetter(1)) #shuffle this list somehow...
-
-	# find the most common word's prevalence (i.e. the last word on this list)
-	maximum_prevalence = added_word_counts[it[len(it)-1][0]]
-
-	for w in it:
-		smallest_word = w[0]
-
-		# if you have already added all the examples from this word - continue
-		if known_word_counts[smallest_word] == added_word_counts[smallest_word]:
-			continue
-
-		# if the least common word has reached the desired count - we're done!
-		if not added_word_counts[smallest_word] < minimum_desired_count:
-			looping = False
-			break
-
-		# find the best unadded sentence for this word - see below
-		best_sentence = None
-		best_sentence_score = 0
-
-		# loop through sentences containing this word
-		for sentence in smallest_word.related:
-			if sentence in added_sentences:
-				continue
-				
-			# base score is number of unique words
-			sentence_score = len(sentence.related) 
-			
-			# score bonus points for how rare each word is
-			for ww in sentence.related:
-				sentence_score += maximum_prevalence - added_word_counts[ww]
-
-			# check if best sentence for this word
-			if sentence_score > best_sentence_score:
-				best_sentence = sentence
-				best_sentence_score = sentence_score
-
-		# if no suitable sentences found - go to next smallest word
-		if not sentence:
-			continue
-
-		# add this sentence - register as "added"
-		covering.append(best_sentence)
-		added_sentences.add(best_sentence)
-
-		# update added_word_counts to include new sentence
-		for ww in best_sentence.related:
-			increment_word(added_word_counts, ww)
-		
-		break # back to main loop
+# FILTER SENTENCE
+sf = Sentence_Filter()
+filtered_sentences = sf.get_filtered_sentences(context)
 
 # output sentence list
-print ("// Learning " +str(len(covering))).encode('utf-8')
-for s in covering:
+print ("// Learning " +str(len(filtered_sentences))).encode('utf-8')
+for s in filtered_sentences:
 	if s.text and s.pinyin and s.english:
 		#print (s.text +"\t" +s.pinyin +"\t" +s.english).encode('utf-8')
 		#print (s.english +"\t" +s.pinyin +"\t" +s.text).encode('utf-8')
@@ -121,7 +40,7 @@ for s in covering:
 
 #known_word_counts = dict()
 
-#for s in covering:
+#for s in filtered_sentences:
 #	example = s.text
 #
 #	for w in words:
